requirements.txt
Django==1.10.1
django-cors-headers==1.1.0
djangorestframework==3.4.6
